@@ -9,21 +9,49 @@
             "email": "devel@keboola.com"
         }
     ],
-    "require-dev": {
-        "squizlabs/php_codesniffer": "^3.0",
-        "phpunit/phpunit": "^5.7"
+    "autoload": {
+        "psr-4": {
+            "Keboola\\Gelf\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Keboola\\Gelf\\Tests\\": "tests/"
+        }
     },
     "require": {
+        "php": "^7.4",
         "react/socket": "^0.8",
         "react/datagram": "^1.2",
         "graylog2/gelf-php": "^1.5",
-        "symfony/process": "^2.8|^4.1",
+        "symfony/process": "^4.1|^5.4",
         "ext-json": "*"
     },
-    "autoload": {
-        "psr-4": {
-            "Keboola\\Gelf\\": "src/",
-            "Keboola\\Gelf\\Tests\\": "tests/"
+    "require-dev": {
+        "keboola/coding-standard": ">=7.0.2",
+        "phpstan/phpstan": "^1.5",
+        "phpstan/phpstan-phpunit": "^1.0.0",
+        "phpunit/phpunit": "^9.3",
+        "ext-sockets": "*",
+        "ext-zlib": "*"
+    },
+    "scripts": {
+        "tests": "phpunit --coverage-clover build/logs/clover.xml --coverage-xml=build/logs/coverage-xml --log-junit=build/logs/phpunit.junit.xml",
+        "phpstan": "phpstan analyse --no-progress -c phpstan.neon",
+        "phpcs": "phpcs --extensions=php src tests",
+        "build": [
+            "@tests"
+        ],
+        "ci": [
+            "@composer validate --no-check-all --strict",
+            "@build"
+        ]
+    },
+    "config": {
+        "sort-packages": true,
+        "process-timeout": 3600,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
         }
     }
 }
