@@ -1,10 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Keboola\Gelf\Tests;
 
-abstract class AbstractGelfTest extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+abstract class AbstractGelfTest extends TestCase
 {
-    protected function checkResults(array $events)
+    protected function checkResults(array $events): void
     {
         $timestamps = [];
         $hosts = [];
@@ -13,7 +17,7 @@ abstract class AbstractGelfTest extends \PHPUnit_Framework_TestCase
         $exception = '';
         foreach ($events as $event) {
             self::assertArrayHasKey('timestamp', $event);
-            self::assertTrue(!empty($event['host']) || ($event['short_message'] == 'No host'));
+            self::assertTrue(!empty($event['host']) || ($event['short_message'] === 'No host'));
             $timestamps[] = $event['timestamp'];
             $hosts[] = empty($event['host']) ? 'empty' : $event['host'];
             if (!empty($event['file'])) {
@@ -37,8 +41,8 @@ abstract class AbstractGelfTest extends \PHPUnit_Framework_TestCase
                 'level' => 1,
                 'timestamp' => $timestamps[1],
                 '_structure' => [
-                    'data' => [0, 1]
-                ]
+                    'data' => [0, 1],
+                ],
             ],
             2 => [
                 'version' => '1.0',
@@ -72,7 +76,7 @@ abstract class AbstractGelfTest extends \PHPUnit_Framework_TestCase
                     'with' => [
                         'several' => 'nested',
                         0 => 'levels',
-                    ]
+                    ],
                 ],
             ],
             5 => [
@@ -102,7 +106,7 @@ abstract class AbstractGelfTest extends \PHPUnit_Framework_TestCase
                 'short_message' => 'Second message',
                 'level' => 7,
                 'timestamp' => $timestamps[8],
-            ]
+            ],
         ], $events);
     }
 }
