@@ -1,7 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Keboola\Gelf\Exception;
 
-class InitException extends \RuntimeException
+use RuntimeException;
+
+class InitException extends RuntimeException
 {
 }
