@@ -1,8 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
+use Psr\Log\LogLevel;
+
 require_once __DIR__ . DIRECTORY_SEPARATOR . '../../vendor/autoload.php';
 
-$server = "127.0.0.1";
+$server = '127.0.0.1';
 $port = 12202;
 $transport = new Gelf\Transport\HttpTransport($server, $port);
 
@@ -11,27 +15,27 @@ $publisher->addTransport($transport);
 
 $logger = new Gelf\Logger($publisher);
 
-$logger->debug("A debug message.");
-$logger->alert("An alert message", ['structure' => ['data' => [0, 1]]]);
+$logger->debug('A debug message.');
+$logger->alert('An alert message', ['structure' => ['data' => [0, 1]]]);
 
 try {
-    throw new Exception("Test exception");
-} catch (Exception $e) {
-    $logger->emergency("Exception example", ['exception' => $e]);
+    throw new Exception('Test exception');
+} catch (Throwable $e) {
+    $logger->emergency('Exception example', ['exception' => $e]);
 }
 
 $message = new Gelf\Message();
-$message->setShortMessage("Structured message")
-    ->setLevel(\Psr\Log\LogLevel::ALERT)
-    ->setFullMessage("There was a foo in bar")
-    ->setFacility("example-facility")
+$message->setShortMessage('Structured message')
+    ->setLevel(LogLevel::ALERT)
+    ->setFullMessage('There was a foo in bar')
+    ->setFacility('example-facility')
     ->setAdditional('foo', 'bar')
     ->setAdditional('bar', 'baz')
 ;
 $publisher->publish($message);
 
-$logger->warning("A warning message.", ['structure' => ['with' => ['several' => 'nested', 'levels']]]);
-$logger->info(file_get_contents(__DIR__ . DIRECTORY_SEPARATOR . "bacon.txt"));
+$logger->warning('A warning message.', ['structure' => ['with' => ['several' => 'nested', 'levels']]]);
+$logger->info(file_get_contents(__DIR__ . DIRECTORY_SEPARATOR . 'bacon.txt'));
 
 // manually create a socket to send some garbage in
 $socket = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
