@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Keboola\Gelf\Tests;
 
 use Keboola\Gelf\ServerFactory;
@@ -7,13 +9,13 @@ use Symfony\Component\Process\Process;
 
 class HttpStreamServerTest extends AbstractGelfTest
 {
-    public function testServer()
+    public function testServer(): void
     {
         $server = ServerFactory::createServer(ServerFactory::SERVER_HTTP);
         $events = [];
         $fails = [];
         $process = new Process(
-            'php ' . __DIR__ . DIRECTORY_SEPARATOR . 'Clients' . DIRECTORY_SEPARATOR . 'HttpClient.php'
+            ['php',  __DIR__ . DIRECTORY_SEPARATOR . 'Clients' . DIRECTORY_SEPARATOR . 'HttpClient.php']
         );
         $server->start(
             12202,
